# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Namma Kasa API",
    "version": "0.1.0",
    "description": "Live tracking of door-to-door waste collection. Errors are RFC 9457 problem+json."
  },
  "servers": [
    {
      "url": "http://localhost:4000/v1",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/admin/complaints": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ward complaint queue, with the collection evidence for each",
        "parameters": [
          {
            "in": "query",
            "name": "wardId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complaints",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminComplaint"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/complaints/{id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Move a complaint through its status workflow",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "open",
                      "in_review",
                      "resolved",
                      "rejected"
                    ]
                  },
                  "resolutionNote": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Complaint"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/resident/complaints": {
      "get": {
        "tags": [
          "resident"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Complaints raised by this household",
        "responses": {
          "200": {
            "description": "Complaints",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Complaint"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resident"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Raise a complaint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string",
                    "enum": [
                      "missed_pickup",
                      "late",
                      "behavior",
                      "segregation",
                      "other"
                    ]
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "mediaUrls": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri"
                    },
                    "maxItems": 3,
                    "default": []
                  }
                },
                "required": [
                  "category"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Complaint"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/resident/ratings": {
      "post": {
        "tags": [
          "resident"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rate today's collection, once per day",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "stars": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "comment": {
                    "type": "string",
                    "maxLength": 1000
                  }
                },
                "required": [
                  "stars"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rating"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/me": {
      "delete": {
        "tags": [
          "account"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request account deletion; personal data is erased after 30 days",
        "responses": {
          "202": {
            "description": "Scheduled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "erasesAfter": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "note": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "erasesAfter",
                    "note"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/me/retention-policy": {
      "get": {
        "tags": [
          "account"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "How long each kind of data is kept",
        "responses": {
          "200": {
            "description": "Policy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deletionGraceDays": {
                      "type": "integer"
                    },
                    "mediaRetentionDays": {
                      "type": "integer"
                    },
                    "pingRetentionDays": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "deletionGraceDays",
                    "mediaRetentionDays",
                    "pingRetentionDays"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "default": "about:blank"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "Complaint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "type": "string",
            "enum": [
              "missed_pickup",
              "late",
              "behavior",
              "segregation",
              "other"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "mediaUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_review",
              "resolved",
              "rejected"
            ]
          },
          "routeId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "wardId": {
            "type": "string",
            "format": "uuid"
          },
          "slaDueAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "resolutionNote": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "history": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "fromStatus": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "enum": [
                    "open",
                    "in_review",
                    "resolved",
                    "rejected",
                    null
                  ]
                },
                "toStatus": {
                  "type": "string",
                  "enum": [
                    "open",
                    "in_review",
                    "resolved",
                    "rejected"
                  ]
                },
                "note": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "at": {
                  "type": "string"
                }
              },
              "required": [
                "fromStatus",
                "toStatus",
                "note",
                "at"
              ]
            }
          }
        },
        "required": [
          "id",
          "category",
          "description",
          "mediaUrls",
          "status",
          "routeId",
          "wardId",
          "slaDueAt",
          "resolutionNote",
          "createdAt",
          "history"
        ]
      },
      "AdminComplaint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "type": "string",
            "enum": [
              "missed_pickup",
              "late",
              "behavior",
              "segregation",
              "other"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "mediaUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_review",
              "resolved",
              "rejected"
            ]
          },
          "routeId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "wardId": {
            "type": "string",
            "format": "uuid"
          },
          "slaDueAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "resolutionNote": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "history": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "fromStatus": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "enum": [
                    "open",
                    "in_review",
                    "resolved",
                    "rejected",
                    null
                  ]
                },
                "toStatus": {
                  "type": "string",
                  "enum": [
                    "open",
                    "in_review",
                    "resolved",
                    "rejected"
                  ]
                },
                "note": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "at": {
                  "type": "string"
                }
              },
              "required": [
                "fromStatus",
                "toStatus",
                "note",
                "at"
              ]
            }
          },
          "household": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "fullName": {
                "type": "string"
              },
              "addressLine": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "fullName",
              "addressLine"
            ]
          },
          "evidence": {
            "type": "object",
            "properties": {
              "servedOnComplaintDay": {
                "type": "boolean"
              },
              "lastCollectedAt": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "servedOnComplaintDay",
              "lastCollectedAt"
            ]
          },
          "slaBreached": {
            "type": "boolean"
          },
          "slaEscalated": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "category",
          "description",
          "mediaUrls",
          "status",
          "routeId",
          "wardId",
          "slaDueAt",
          "resolutionNote",
          "createdAt",
          "history",
          "household",
          "evidence",
          "slaBreached",
          "slaEscalated"
        ]
      },
      "Rating": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "stars": {
            "type": "integer"
          },
          "comment": {
            "type": [
              "string",
              "null"
            ]
          },
          "collectionDate": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "stars",
          "comment",
          "collectionDate",
          "createdAt"
        ]
      },
...
    },
    "responses": {
      "Problem": {
        "description": "RFC 9457 problem document",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    }
  }
}